Added loader by 5 comment items
